Eliminated the need to specify Binding Flags
Enhanced the extension methods so that the Binding Flags no longer need
to b explicitly stated.  It will automatically search all private,
protected, internal, public, static, instance members.

NOTE: It is forced to only look at the child class and not at the parent
class to prevent finding inherited duplicates.  This is probably the
most common case.  The other case can still be handled if the binding
flags are explicitly stated for that particular call.